move normalizers da raiz para camada transform
