+++ etl/pipeline/standardizer.py
@@ -9,7 +9,7 @@ from etl.extractors import (
     normalize_identifiers,
     rebuild_abstract_from_inverted_index,
 )
-from etl.normalizers import (
+from etl.transform.normalizers import (
     as_list,
     int_or_none,
     normalize_keywords,
